requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
flickrapi==2.4.0
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
oauthlib==2.1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.11.0
twython==3.7.0
urllib3==1.23
